+++ lib/widgets/ny_base_state.dart
@@ -280,7 +280,7 @@ abstract class NyBaseState<T extends StatefulWidget> extends State<T> {
             }
             return Nylo.appLoader();
           }
-        case LoadingStyleType.skeltonizer:
+        case LoadingStyleType.skeletonizer:
           {
             if (loadingStyle.child != null) {
               return Skeletonizer(
@@ -826,38 +826,48 @@ abstract class NyBaseState<T extends StatefulWidget> extends State<T> {
 }
 
 /// The [LoadingStyleType] enum is used to determine the type of loading widget
-enum LoadingStyleType { normal, skeltonizer, none }
+enum LoadingStyleType { normal, skeletonizer, none }
+
+enum SkeletonizerEffect { shimmer, leaf, pulse }
 
 /// The [LoadingStyle] class is used to determine the type of loading widget
 class LoadingStyle {
   final LoadingStyleType type;
   final Widget? child;
+  final SkeletonizerEffect? skeletonizerEffect;
 
-  /// Construct a [LoadingStyle.skletonizer] widget
+  /// Construct a [LoadingStyle.skeletonizer] widget
   /// Provide a [widget] to display a custom loading widget.
-  /// By default, the [LoadingStyle.skletonizer] widget will attempt to
-  /// render your widget with a skeltonizer loader.
-  LoadingStyle.skeletonizer({this.child}) : type = LoadingStyleType.skeltonizer;
+  /// By default, the [LoadingStyle.skeletonizer] widget will attempt to
+  /// render your widget with a skeletonizer loader.
+  LoadingStyle.skeletonizer({this.child, SkeletonizerEffect? effect})
+      : type = LoadingStyleType.skeletonizer,
+        skeletonizerEffect = effect;
 
   /// Construct a [LoadingStyle.normal] widget
   /// Provide a [widget] to display a custom loading widget.
   /// By default, the [LoadingStyle.normal] widget will display your Nylo loader.
-  LoadingStyle.normal({this.child}) : type = LoadingStyleType.normal;
+  LoadingStyle.normal({this.child})
+      : type = LoadingStyleType.normal,
+        skeletonizerEffect = null;
 
   /// Construct a [LoadingStyle.none] widget
   /// This will not display any loading widget.
   LoadingStyle.none()
       : type = LoadingStyleType.none,
+        skeletonizerEffect = null,
         child = null;
 
-  LoadingStyle({required this.type, this.child});
+  LoadingStyle({this.child})
+      : type = LoadingStyleType.normal,
+        skeletonizerEffect = null;
 
   /// Render the loading widget
   Widget render() {
     switch (type) {
       case LoadingStyleType.normal:
         return child ?? Nylo.appLoader();
-      case LoadingStyleType.skeltonizer:
+      case LoadingStyleType.skeletonizer:
         return Skeletonizer(
           enabled: true,
           child: child ?? Nylo.appLoader(),